requirements.txt
# Essential dependencies for Google Gemini API project
google-generativeai==0.8.3
python-dotenv==1.0.0
streamlit==1.32.0
